start working on new register script for apply.bethel.edu that uses WSAPI
